+++ web/src/ee/features/in-app-agent/constants.ts
@@ -1,0 +1,2 @@
+// Avoid renaming this as the FE is aware of this when rendering, renaming it would cause history issues
+export const IN_APP_AGENT_REDIRECT_TOOL_NAME = "langfuse_proposeRedirect";
